+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+- Added Metrics.firstCycle.
+- Exposed convertModelToDirectedGraph in the public API.
+- Added more dartdoc comments to the public API.
+- Made subgraph ids relative to rootDir instead of rootDir.parent.
+
 ## 1.0.0
 
 - Initial version.